# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Dream Journal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Emotion intensities for a dream (each 0-10)
 */
export interface Emotions {
  /**
     * @minimum 0
     * @maximum 10
     */
  angry: number;
  /**
     * @minimum 0
     * @maximum 10
     */
  sad: number;
  /**
     * @minimum 0
     * @maximum 10
     */
  disgust: number;
  /**
     * @minimum 0
     * @maximum 10
     */
  happy: number;
  /**
     * @minimum 0
     * @maximum 10
     */
  peaceful: number;
  /**
     * @minimum 0
     * @maximum 10
     */
  neutral: number;
}

export interface Dream {
  id: number;
  /** Date of the dream (YYYY-MM-DD) */
  date: string;
  title: string;
  description: string;
  emotions: Emotions;
  /**
     * How lucid the dream was (0=not aware, 10=fully lucid)
     * @minimum 0
     * @maximum 10
     */
  lucidity: number;
  /**
     * How vivid/clear the dream felt (0=foggy, 10=crystal clear)
     * @minimum 0
     * @maximum 10
     */
  clarity: number;
  /**
     * Nightmare intensity (0=pleasant, 10=extreme nightmare)
     * @minimum 0
     * @maximum 10
     */
  nightmareFactor: number;
  /**
     * ISO timestamp when the dream was archived, or null if active
     * @nullable
     */
  archivedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface DreamInput {
  /** Date of the dream (YYYY-MM-DD) */
  date: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  emotions: Emotions;
  /**
     * @minimum 0
     * @maximum 10
     */
  lucidity: number;
  /**
     * @minimum 0
     * @maximum 10
     */
  clarity: number;
  /**
     * @minimum 0
     * @maximum 10
     */
  nightmareFactor: number;
}

export interface DreamUpdate {
  date?: string;
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  description?: string;
  emotions?: Emotions;
  /**
     * @minimum 0
     * @maximum 10
     */
  lucidity?: number;
  /**
     * @minimum 0
     * @maximum 10
     */
  clarity?: number;
  /**
     * @minimum 0
     * @maximum 10
     */
  nightmareFactor?: number;
}

export interface DreamStats {
  totalEntries: number;
  /** Number of consecutive days with a dream entry */
  currentStreak: number;
  /** All-time longest consecutive daily entry streak */
  longestStreak: number;
  avgLucidity: number;
  avgClarity: number;
  avgNightmareFactor: number;
  /** Average word count of dream descriptions */
  avgEntryLength: number;
  /** Fraction of entries with nightmareFactor >= 7 (0.0 - 1.0) */
  nightmareFrequency: number;
  avgEmotions: Emotions;
}