# lib/api-zod/src/generated/types/dream.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Dream Journal API
 * OpenAPI spec version: 0.1.0
 */
import type { Emotions } from './emotions';

export interface Dream {
  id: number;
  /** Date of the dream (YYYY-MM-DD) */
  date: string;
  title: string;
  description: string;
  emotions: Emotions;
  /**
     * How lucid the dream was (0=not aware, 10=fully lucid)
     * @minimum 0
     * @maximum 10
     */
  lucidity: number;
  /**
     * How vivid/clear the dream felt (0=foggy, 10=crystal clear)
     * @minimum 0
     * @maximum 10
     */
  clarity: number;
  /**
     * Nightmare intensity (0=pleasant, 10=extreme nightmare)
     * @minimum 0
     * @maximum 10
     */
  nightmareFactor: number;
  /**
     * ISO timestamp when the dream was archived, or null if active
     * @nullable
     */
  archivedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}